Use Room for filelist table reads

+++ app/src/main/java/com/nextcloud/client/database/entity/FileEntity.kt
@@ -31,7 +31,7 @@ import com.owncloud.android.db.ProviderMeta.ProviderTableMeta
 data class FileEntity(
     @PrimaryKey(autoGenerate = true)
     @ColumnInfo(name = ProviderTableMeta._ID)
-    val id: Int?,
+    val id: Long?,
     @ColumnInfo(name = ProviderTableMeta.FILE_NAME)
     val name: String?,
     @ColumnInfo(name = ProviderTableMeta.FILE_ENCRYPTED_NAME)
@@ -41,25 +41,25 @@ data class FileEntity(
     @ColumnInfo(name = ProviderTableMeta.FILE_PATH_DECRYPTED)
     val pathDecrypted: String?,
     @ColumnInfo(name = ProviderTableMeta.FILE_PARENT)
-    val parent: Int?,
+    val parent: Long?,
     @ColumnInfo(name = ProviderTableMeta.FILE_CREATION)
-    val creation: Int?,
+    val creation: Long?,
     @ColumnInfo(name = ProviderTableMeta.FILE_MODIFIED)
-    val modified: Int?,
+    val modified: Long?,
     @ColumnInfo(name = ProviderTableMeta.FILE_CONTENT_TYPE)
     val contentType: String?,
     @ColumnInfo(name = ProviderTableMeta.FILE_CONTENT_LENGTH)
-    val contentLength: Int?,
+    val contentLength: Long?,
     @ColumnInfo(name = ProviderTableMeta.FILE_STORAGE_PATH)
     val storagePath: String?,
     @ColumnInfo(name = ProviderTableMeta.FILE_ACCOUNT_OWNER)
     val accountOwner: String?,
     @ColumnInfo(name = ProviderTableMeta.FILE_LAST_SYNC_DATE)
-    val lastSyncDate: Int?,
+    val lastSyncDate: Long?,
     @ColumnInfo(name = ProviderTableMeta.FILE_LAST_SYNC_DATE_FOR_DATA)
-    val lastSyncDateForData: Int?,
+    val lastSyncDateForData: Long?,
     @ColumnInfo(name = ProviderTableMeta.FILE_MODIFIED_AT_LAST_SYNC_FOR_DATA)
-    val modifiedAtLastSyncForData: Int?,
+    val modifiedAtLastSyncForData: Long?,
     @ColumnInfo(name = ProviderTableMeta.FILE_ETAG)
     val etag: String?,
     @ColumnInfo(name = ProviderTableMeta.FILE_ETAG_ON_SERVER)
@@ -111,7 +111,7 @@ data class FileEntity(
     @ColumnInfo(name = ProviderTableMeta.FILE_LOCK_OWNER_EDITOR)
     val lockOwnerEditor: String?,
     @ColumnInfo(name = ProviderTableMeta.FILE_LOCK_TIMESTAMP)
-    val lockTimestamp: Int?,
+    val lockTimestamp: Long?,
     @ColumnInfo(name = ProviderTableMeta.FILE_LOCK_TIMEOUT)
     val lockTimeout: Int?,
     @ColumnInfo(name = ProviderTableMeta.FILE_LOCK_TOKEN)
